not discard multiple shares
When preparing the data for the incidence matrix, we have the choice to disregard multiple shares (i.e., different content_id) of the same object by the same user. This was the current implementation, where deduplication were done by considering user_id and object_id. We now added the content_id to consider pairs of objects and contents when deduplicating. Using this approach, we keep into account multiple shares of the same object by users. This change is applied conditionally to intent = "users", as the implication for the other types of detection should be considered.