simplified graph creation
Simplified the creation of the coordinated network, using the aggregated result datatable. The steps relating to the incidence matrix have consequently been eliminated in this version. Furthermore, the construction of the matrix increased the number of edges unnecessarily, as it connected all the users who had shared the same object_id, while only those who shared the same object_id in the same time_window need to be connected.