edge_simmetry_score
The graph now incorporates the contribution of each node within the pair to the pair's edge weight, specifically, the number of shared content_id that contribute to the edge weight. Additionally, an edge_symmetry_score is included, which equals 1 in cases of equal contributions from both users and approaches 0 as the contributions become increasingly unequal. This score, along with the value of contributions, can be utilized for further filtering or examining cases where the score is particularly low. Working with an undirected graph, it is plausible that the activity of highly active users disproportionately affects the weight of edges connecting them to less active users. For instance, if user A shares the same objects (object_id) 100 times, and user B shares the same object only once, but within a time frame that matches the time_window defined in the parameter for all of user A's 100 shares, then the edge weight between A and B will be 100, although this weight is almost entirely influenced by the hyperactivity of user A. The edge_symmetry_score, along with the counts of shares by each user (n_content_id and n_content_id_y), allows for monitoring and controlling this phenomenon