Merge new simulation feature into main branch (#28)
* add function for simulation
* Update simulate_data.R

The function now includes a parameter dataset_size that allows the user to specify the size of the final dataset. Additionally, there is a parameter prop_coord_size to control the proportion of coordinated content relative to the total dataset_size. The function automatically estimates the lambda parameters for both coordinated and non-coordinated Poisson distributions and uses an iterative process to achieve the desired dataset size.

A parameter approx_size has been added, allowing control over the approximate size of the dataset by setting the lambdas appropriately, considering the number of requested accounts. The user can also directly specify lambdas for coordinated and non-coordinated networks, which disables the approx_size parameter and returns a warning.

* validate input parameters
* documentation updated

The documentation for the function has been updated to include the new parameters and their description

---------

Co-authored-by: Nicola Righetti